update example configs
