+++ configs/config-example.json
@@ -4,6 +4,8 @@
     "zookeeper"
   ],
   "commonConfig": {
+    "httpPort": 8080,
+    "docIndex": "",
     "dxApiBasePath": "",
     "dxAuthBasePath": "",
     "@context": "",
@@ -53,7 +55,6 @@
       "catAdmin": "",
       "verticleInstances": 2,
       "ip": "",
-      "httpPort": 8080,
       "ssl": false,
       "host": "",
       "databaseIP": "",
@@ -88,13 +89,19 @@
     {
       "id": "iudx.catalogue.server.rating.RatingVerticle",
       "verticleInstances": 1,
+      "ratingDatabaseIP": "",
+      "ratingDatabasePort": 5433,
+      "ratingDatabaseName": "",
+      "ratingDatabaseUserName": "",
+      "ratingDatabasePassword": "",
+      "ratingPoolSize": 25,
       "ratingExchangeName": "",
       "rsAuditingTableName": "",
       "ratingIndex": "",
       "minReadNumber": 100
     },
     {
-      "id": "iudx.catalogue.server.databroker.rabbitmq.RabbitMQVerticle",
+      "id": "iudx.catalogue.server.databroker.RabbitMQVerticle",
       "verticleInstances": 1,
       "dataBrokerIP": "",
       "dataBrokerPort": 1234,
