+++ configs/config-test-examples.json
@@ -56,10 +56,15 @@
       "databasePort": 1234,
       "databasePassword": "",
       "databaseUser": "",
-      "docIndex": ""
+      "docIndex": "",
+      "@context": ""
     },
     {
       "id": "iudx.catalogue.server.apiserver.ApiServerVerticle",
+      "optionalModules": [
+        "iudx.catalogue.server.geocoding.GeocodingVerticle",
+        "iudx.catalogue.server.nlp.NLPSearchVerticle"
+      ],
       "keystorePath": "",
       "keystorePassword": "",
       "token":"",
@@ -70,6 +75,7 @@
       "httpPort": 1234,
       "ssl": false,
       "host": "",
+      "dxApiBasePath": "",
       "isUACInstance": false,
       "databaseIP": "",
       "docIndex": "",
@@ -111,7 +117,9 @@
       "ratingPoolSize": 25,
       "ratingExchangeName": "",
       "rsAuditingTableName": "",
-      "minReadNumber": 100
+      "minReadNumber": 100,
+      "docIndex": "",
+      "ratingIndex": ""
     },
     {
       "id": "iudx.catalogue.server.databroker.RabbitMQVerticle",
@@ -133,6 +141,9 @@
       "verticleInstances": 1,
       "databaseTable": "",
       "catSummaryTable" : "",
+      "docIndex": "",
+      "mlayerInstanceIndex": "",
+      "mlayerDomainIndex": "",
       "excluded_ids": []
     },
     {
