+++ src/main/java/iudx/catalogue/server/apiserver/item/model/Item.java
